Measured boot: Limit output param sizes to buffer size
The measured boot read API receives requested output parameter sizes
from the caller via the iovec out_size and passes them to the internal
API as the buffer size for the parameters.

However, the caller could send a requested size much larger than size of
the local buffer. In theory, that could lead to the internal API writing
past the end of the local buffer, but it cannot happen in practice
because the local buffer is large enough for the largest response.

Even so, it is good practice to sanitise parameters from the untrusted
caller to guarantee that buffer overrun will not happen, so this patch
limits the sizes to the buffer size.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I2fa260297f4c6088a8cdf4f5e940cdd3d00ae9e8