Measured boot: Format version string without stdlib
The previous commit 884371e added support for formatting the version
string in the measured boot partition using snprintf. Unfortunately,
that introduced heap usage, which is not supported for TF-M partitions,
and causes build errors in release builds.

This patch solves the issue by manually formatting the string instead of
using stdlib functions. As the string is quite simple, this also has the
advantage of reducing code size.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I0c6a16f55929974d4d5ad59a1b20cb773a0687b2