Build: Add dependencies for the manifest tool
This patch adds dependencies for the manifest tool so that any change on
the files the manifest tool consumes will trigger the manifest tool to
re-run thus make the changes take effect.

This patch also removes the unnecessary link to psa_interface as it has
been publicly linked by tfm_sprt.

Signed-off-by: Kevin Peng <kevin.peng@arm.com>
Change-Id: I3170a86e50dfdc946f56ff451ae6fa1436e2aeae