DPE: Fix NS tests
With split build, the dpe_api_ns interface library does not work for
linking the NS tests to the NS APIs because the library is only part of
the S build. This patch replaces the interface library with install of
all required NS files and directly building them into the NS tests.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: Ic42ca8eb258f6d08b4c70bf6b4a2b9c3b9170359