DPE: Support host build
If `-DHOST_BUILD=ON` is added to the CMake command line
then the DPE library will be built as a stand-alone
Linux program. Not tested on Windows.

It allows the rapid development, build and test cycles
of the DPE library in a native environment without
the need to execute the code on FVP.

Change-Id: I4923c2aba9636d96bfe1aca4e7d5b731270382fa
Signed-off-by: Tamas Ban <tamas.ban@arm.com>