Revert "DelegAttest: Fix build errors"
This reverts commit 640630fa81524e813fe193b5f21d9cfc47866d93.

Reason for revert:

The TF-A nightly CI job picks up the main branch. But it uses prebuilt binaries for RSE, so there is no TF-M build. The test suite of delegated attestation and measured boot partitions are linked to the BL31 image to execute the test suite from the AP side.

Since `psa/initial_attestation.h` is generated during the TF-M build, it is unavailable at the TF-A build time. This causes a build error.

We need to look into finding a proper solution until reverting the commit is the best.

Change-Id: I8faa8e26ea5138742976a559366a3e2b62e5f22a
Signed-off-by: Tamas Ban <tamas.ban@arm.com>