RIV: Add RSE Image Verification partition
The RSE Image Verification (RIV) partition provides one service that
can be used to load and verify images during the runtime. The loaded
image is also measured and the measurement can be stored using the
measured_boot partition.

The partition does not prescribe a particular image format: the
verification struct can be populated to support different image formats.
There are example verification structure setups for the MCUBoot image
format, and for the ST signed ELF image format.

Change-Id: I51fa7cac520a353be30469c44425d226d78b52be
Signed-off-by: Bence Balogh <bence.balogh@arm.com>