SCMI: Add support for transport direction
The shared memory with SCP is used for both commands and notifications.
It's mandates to keep into account the direction of such
communications between the agent (A) and the platform (P).

The shared memory must be split into 2 halves so that:
 - A->P channels for commands
 - P->A channels for notifications and delayed responses

That translates to having shared_memory_a2p and shared_memory_p2a
channels within the same shared memory.

Testsuite is also updated.

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I6eb582063f3fb7053adabfbe8424cafe4a4b814f